+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - Add DSN getters ([#540](https://github.com/getsentry/sentry-go/pull/540))
 - Add Span.SetData() ([#542](https://github.com/getsentry/sentry-go/pull/542))
 - Add Span.IsTransaction() ([#543](https://github.com/getsentry/sentry-go/pull/543))
+- Add TraceParentContext structure ([#545](https://github.com/getsentry/sentry-go/pull/545))
+- Add a new SpanOption: `SpanSampled` ([#546](https://github.com/getsentry/sentry-go/pull/546))
 
 ## 0.17.0